refactor Sign up view
